feat: 글쓰기 완료 후 리스트로 이동 및 리스트에 글쓰기 버튼 추가
- /board/register에서 /board/list로 이동시 글쓰기 완료를 알리는 modal창이 뜬다
- list에 register로 이동 할 수 있는 버튼 추가

+++ src/main/java/org/gosang/controller/BoardController.java
@@ -51,7 +51,7 @@ public class BoardController {
 	
 	
 	@GetMapping("/get")
-	public void get(@RequestParam("bno") Long bno, Model model) {
+	public void get(@RequestParam("bno") int bno, Model model) {
 		
 		log.info("/get");
 		model.addAttribute("board", service.get(bno));
@@ -70,7 +70,7 @@ public class BoardController {
 	}
 	
 	@PostMapping("/remove")
-	public String remove(@RequestParam("bno") Long bno, RedirectAttributes rttr) {
+	public String remove(@RequestParam("bno") int bno, RedirectAttributes rttr) {
 		
 		log.info("remove..." + bno);
 		if(service.remove(bno)) {
