+++ src/main/java/org/gosang/domain/BoardVO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class BoardVO {
 
-	private Long bno;
+	private int bno;			// null값 처리를 위해 int가 아닌 Integer 사용 
 	private String title;
 	private String content;
 	private String writer;
