+++ src/main/java/org/gosang/service/BoardService.java
@@ -8,11 +8,11 @@ public interface BoardService {
 	
 	public void register(BoardVO board);
 	
-	public BoardVO get(Long bno);
+	public BoardVO get(int bno);
 	
 	public boolean modify(BoardVO board);
 	
-	public boolean remove(Long bno);
+	public boolean remove(int bno);
 	
 	public List<BoardVO> getList();
 }
